move buildSrc logic into the main stuff
