--- firebase-dataconnect/gradleplugin/buildSrc/build.gradle.kts
@@ -1,29 +1,0 @@
-// Copyright 2024 Google LLC
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//      http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-plugins {
-  alias(libs.plugins.kotlin.jvm)
-  alias(libs.plugins.kotlinx.serialization)
-  alias(libs.plugins.spotless)
-  `kotlin-dsl`
-}
-
-java { toolchain { languageVersion.set(JavaLanguageVersion.of(17)) } }
-
-dependencies {
-  compileOnly(libs.android.gradlePlugin.gradle.api)
-  implementation(gradleKotlinDsl())
-  implementation(libs.kotlinx.serialization.core)
-  implementation(libs.kotlinx.serialization.json)
-}
